CUT: add corosync.conf.example
Issue: LIN1018-130

* Add corosync.conf.example as the template for
  /etc/corosync/corosync.conf to make the cluster
  works as the previous default config doesn't work
  as below output:

  root@host1:/# crm status
  Stack: unknown
  Current DC: NONE
  Last updated: Mon Dec 25 05:30:59 2017
  Last change: Mon Dec 25 05:30:32 2017 by root via cibadmin on host1

  0 nodes configured
  1 resource configured

  Full list of resources:

  WebSite	(ocf::heartbeat:apache):	Stopped

+++ wrlinux-cgl/recipes-cut/cgl-unittest/cgl-unittest/files/scripts/cluster.sh
@@ -64,7 +64,6 @@ clean()
 
 cus_corosync()
 {
-    cp /etc/corosync/corosync.conf.example /etc/corosync/corosync.conf
     check_localeth
  
     # get netmask
@@ -83,7 +82,11 @@ cus_corosync()
     fi
     
     # customize the corosync.conf
-    cp /etc/corosync/corosync.conf.example /etc/corosync/corosync.conf
+    if [ -f /opt/cut/resource/corosync.conf.example ]; then
+        cp /opt/cut/resource/corosync.conf.example /etc/corosync/corosync.conf
+    else
+        cp /etc/corosync/corosync.conf.example /etc/corosync/corosync.conf
+    fi
    
     # update the bindnetaddr item in /etc/corosync/corosync.conf
     sed -i "s/bindnetaddr: .*$/bindnetaddr: $NETADDR1/g" /etc/corosync/corosync.conf
